Bike model.
Previously, I was using a 'system' model that tried to provide a unified
view into any type of bike, irrespective of how it was structured or
what components it was made of.

This turned out to be a PITA and more trouble than it's worth. In
particular, if I needed to reconcile differing data from different
components, it was difficult to know why a value was set. Instead,
it's replaced with a Bike model. It represents exactly the components
that I'm using.